Adds back tests with comments
